+++ src/components/layout/desktop/aside/calendar.vue
@@ -12,8 +12,8 @@
 <script lang="ts">
   import { defineComponent, onMounted } from 'vue'
   import { useArticleCalendarStore } from '/@/store/aggregate'
-  import { getDateFlowRoute } from '/@/transforms/route'
   import { humanDateToYMD, HumanDate } from '/@/transforms/moment'
+  import { getDateFlowRoute } from '/@/transforms/route'
   import Calendar from '/@/components/widget/calendar.vue'
 
   export default defineComponent({